refactored into different packages for repeated and oneshot games
