+++ llm_cooperation/dilemma.py
@@ -11,10 +11,9 @@ from llm_cooperation import (
     Choice,
     Choices,
     Payoffs,
-    RepeatedResults,
     run_and_record_experiment,
-    run_experiment_repeated_game,
 )
+from llm_cooperation.repeated import RepeatedGameResults, run_experiment
 
 SAMPLE_SIZE: int = 30
 NUM_ROUNDS: int = 6
@@ -155,8 +154,8 @@ def compute_freq_pd(choices: List[Choices]) -> float:
     return len([c for c in choices if c.ai == Cooperate]) / len(choices)
 
 
-def run_experiment_pd() -> RepeatedResults:
-    return run_experiment_repeated_game(
+def run_experiment_pd() -> RepeatedGameResults:
+    return run_experiment(
         ai_participants=AI_PARTICIPANTS,
         partner_conditions={
             "unconditional cooperate": strategy_cooperate,
